% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{calculate_penalty}
\alias{calculate_penalty}
\title{Calculate penalties}
\usage{
calculate_penalty(
  timedate,
  deadline = c("13:40:00", "13:45:00", "13:50:00"),
  penalty = c(0.3, 0.6, 1)
)
}
\arguments{
\item{timedate}{vector of timedates}

\item{deadline}{vector of deadlines}

\item{penalty}{vector of penalties (same length as deadline)}
}
\value{
penalties vector
}
\description{
Calculate penalties
}
\details{
Calculate penalties
}
