% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{copy_one_gh_entry}
\alias{copy_one_gh_entry}
\title{Copy one github entry: file or folder}
\usage{
copy_one_gh_entry(from, to, name, remove = c("LICENSE", "README.md"))
}
\arguments{
\item{from}{repository folder}

\item{to}{student target folder}

\item{name}{name for target file/folder}

\item{remove}{file names to exclude}
}
\value{
nothing, just copy
}
\description{
Copy one github entry: file or folder
}
\details{
Copy one github entry: file or folder
}
