% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{create_target_fname}
\alias{create_target_fname}
\title{Create target file/folder name from kr, id, time}
\usage{
create_target_fname(kr, time, id)
}
\arguments{
\item{kr}{integer kr number}

\item{time}{datetime of the submission}

\item{id}{student id}
}
\value{
target file/folder name without extension
}
\description{
Create target file/folder name from kr, id, time
}
\details{
Create target file/folder name from kr, id, time
}
