% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{dir_create}
\alias{dir_create}
\title{Create folder}
\usage{
dir_create(path)
}
\arguments{
\item{path}{folder}
}
\value{
nothing, just create folder
}
\description{
Create folder
}
\details{
Create folder. No warning if the folder already exists.
}
