% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{force_unique_date}
\alias{force_unique_date}
\title{Force unique date}
\usage{
force_unique_date(timedate)
}
\arguments{
\item{timedate}{vector with probably different dates}
}
\value{
timedate vector with unique date forced
}
\description{
Force unique date for a vector of timedates
}
\details{
Force unique date for a vector of timedates.
}
