% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{get_gh_commits}
\alias{get_gh_commits}
\title{Get list of git repository commits}
\usage{
get_gh_commits(path = ".")
}
\arguments{
\item{path}{folder name}
}
\value{
tibble with useful id data
}
\description{
Get list of git repository commits
}
\details{
Get list of git repository commits
}
