% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{get_lms_prefix}
\alias{get_lms_prefix}
\title{Get lms prefix}
\usage{
get_lms_prefix(last_name, first_name)
}
\arguments{
\item{last_name}{last name in cyrillic}

\item{first_name}{first name in cyrillic}
}
\value{
filename prefix in english
}
\description{
Get lms prefix
}
\details{
Get lms prefix
}
