% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{get_lms_unique_submission}
\alias{get_lms_unique_submission}
\title{Get hse lms unique submission}
\usage{
get_lms_unique_submission(
  uploads,
  which = c("first", "last"),
  start = "13:00:00"
)
}
\arguments{
\item{uploads}{tibble with uploads data}

\item{which}{either first or last submission is counted}

\item{start}{submissions before start are ignored and non counted as first submission}
}
\value{
tibble with unique submissions
}
\description{
Get hse lms unique submission
}
\details{
Get hse lms unique submission
}
