% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{get_stud_id_from_repository}
\alias{get_stud_id_from_repository}
\title{Recognise student id using folder content}
\usage{
get_stud_id_from_repository(path = ".")
}
\arguments{
\item{path}{folder name}
}
\value{
student integer id or NA
}
\description{
Recognise student id using folder content
}
\details{
Recognise student id using folder content.
The function searches for the files like krx_xxx.pdf or something close to this.
}
