% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{get_students_data}
\alias{get_students_data}
\title{Get students data from google sheets}
\usage{
get_students_data(path)
}
\arguments{
\item{path}{filename of csv with student data}
}
\value{
tibble with useful id data
}
\description{
Get students data from google sheets
}
\details{
Get students data from google sheets
}
