% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{list_files}
\alias{list_files}
\title{List files in a folder}
\usage{
list_files(path = ".", remove = c("LICENSE", "README.md"))
}
\arguments{
\item{path}{folder name}

\item{remove}{file names to exclude}
}
\value{
character vector with a list of files separated by comma
}
\description{
List files in a folder
}
\details{
List files in a folder
}
