% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{parse_lms_filenames}
\alias{parse_lms_filenames}
\title{Parse lms filenames}
\usage{
parse_lms_filenames(path = ".")
}
\arguments{
\item{path}{folder name}
}
\value{
tibble with useful id data
}
\description{
Parse lms filenames
}
\details{
Parse lms filenames
}
