% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{separate_lms_grain_chaff}
\alias{separate_lms_grain_chaff}
\title{Separate grain from chaffs}
\usage{
separate_lms_grain_chaff(uploads)
}
\arguments{
\item{uploads}{tibble with uploads data}
}
\value{
tibble with good and bad guys
}
\description{
The good, the bad and the ugly :)
}
\details{
Just for information: check whether the submission was unique and with pdf file.
}
