% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_students_data.R
\name{translit}
\alias{translit}
\title{Transliterate cyrillic text}
\usage{
translit(x)
}
\arguments{
\item{x}{cyrillic text}
}
\value{
transliterated text
}
\description{
Transliterate cyrillic text
}
\details{
Transliterate cyrillic text
}
